requirements.txt
flask
flask-cors
pandas 
python-dotenv
torch
kaggle
ipykernel
transformers
accelerate
ipywidgets
sentence_transformers
faiss-cpu
langchain
flash-attn
huggingface-hub
kernels
safetensors
openai-harmony

# i also had to instal  NVIDIA CUDA Compiler driver which is included as part of the CUDA Toolkit
# nvcc --version

# torch 
# pip3 install torch torchvision torchaudio